% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snupassr.R
\name{get_lpass}
\alias{get_lpass}
\title{Get an entry from lastpass via the CLI. Mostly not used
directly. Mostly.}
\usage{
get_lpass(flags, site_entry = "opendata.socrata.com")
}
\arguments{
\item{flags}{are passed to lpass to specify what should be shown}

\item{site_entry}{which entry in lastpass should be shown. Defaults
to opendata.socrata.com.}
}
\value{
a character string
}
\description{
Get an entry from lastpass via the CLI. Mostly not used
directly. Mostly.
}
\examples{
get_lpass("--username", "example.com")
}
