% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snupassr.R
\name{get_pass}
\alias{get_pass}
\title{Get the password field from a lastpass entry.}
\usage{
get_pass(...)
}
\arguments{
\item{site_entry}{which entry in lastpass should be shown. Defaults
to opendata.socrata.com.}
}
\description{
Get the password field from a lastpass entry.
}
