% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snupassr.R
\name{get_token}
\alias{get_token}
\title{Get the "pubtoken" custom field from a lastpass entry.}
\usage{
get_token(...)
}
\arguments{
\item{site_entry}{which entry in lastpass should be shown. Defaults
to opendata.socrata.com.}
}
\description{
Get the "pubtoken" custom field from a lastpass entry.
}
