% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snupassr.R
\name{get_user}
\alias{get_user}
\title{Get the user field from a lastpass entry.}
\usage{
get_user(...)
}
\arguments{
\item{site_entry}{which entry in lastpass should be shown. Defaults
to opendata.socrata.com.}
}
\description{
Get the user field from a lastpass entry.
}
